+++ m0/Libft/ft_strlen.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
-	size_t	len;
+	int	len;
 
 	len = 0;
 	while (str[len])
